added settlement gpu render
