+++ app/Assets/Prefabs/World/SettlementChunk.prefab
@@ -293,7 +293,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7381509102365278593}
-  - component: {fileID: 7653363813633463180}
+  - component: {fileID: 4537980436543227478}
   m_Layer: 0
   m_Name: Ground
   m_TagString: Untagged
@@ -316,7 +316,7 @@ Transform:
   m_Children: []
   m_Father: {fileID: 2209581637102914891}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &7653363813633463180
+--- !u!114 &4537980436543227478
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -325,15 +325,15 @@ MonoBehaviour:
   m_GameObject: {fileID: 4632873142431337680}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2a309a04b03364733bdaf4d2ddc19269, type: 3}
+  m_Script: {fileID: 11500000, guid: 964db7dd93de46b4bdc89d6ca8f5238b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  tilePrefab: {fileID: 1979290889500462, guid: 11995f7d008f146f3a7b9e36db166d5d, type: 3}
-  occupiedTilePrefab: {fileID: 1149535237560096, guid: 638c0a484186943949dddab0d30ab59b,
-    type: 3}
-  surroundingTile: {fileID: 9087089482889845678, guid: 45b79ad5b5b554a67b3e3615a34e0492,
-    type: 3}
-  scaleFactor: 1
+  tileMeshes:
+  - {fileID: -7774794935839024385, guid: 82eb98bff98fe4bfab6a4cf3dcb00879, type: 3}
+  - {fileID: 4300000, guid: c90c72c15c95779418a6c2e7049b777c, type: 3}
+  - {fileID: -1187762388698719382, guid: 6af879b6bb6464fefade050fd58e369f, type: 3}
+  instancedMat: {fileID: 2100000, guid: ab08612e63d71e8419000749e3c68d16, type: 2}
+  scaleFactor: 0.95
 --- !u!1 &5356343550776609569
 GameObject:
   m_ObjectHideFlags: 0
