+++ app/Assets/Scripts/View/Building/DisplayBuildings.cs
@@ -12,7 +12,7 @@ namespace View.Building
         [SerializeField] private BuildingPreview prefab;
 
 
-        public void SetData(global::Settlement.Accounts.Settlement value)
+        public void SetData(Settlement.Accounts.Settlement value, Vector2Int offset)
         {
             if (value == null)
                 return;
