+++ app/Assets/Scripts/View/Building/DisplayWorkers.cs
@@ -7,7 +7,7 @@ namespace View.Building
     {
         [SerializeField] private Worker workerPrefab;
 
-        public void SetData(global::Settlement.Accounts.Settlement value)
+        public void SetData(Settlement.Accounts.Settlement value, Vector2Int offset)
         {
             if (value == null)
                 return;
