+++ app/Assets/Scripts/View/Building/IDisplaySettlementData.cs
@@ -1,7 +1,9 @@
+using UnityEngine;
+
 namespace View.Building
 {
     public interface IDisplaySettlementData
     {
-        void SetData(global::Settlement.Accounts.Settlement value);
+        void SetData(Settlement.Accounts.Settlement value, Vector2Int offset);
     }
 }
